fix: refresh resets recruit state
