+++ lib/presentation/board/recruit/list/view_models/recruit_list_view_model.dart
@@ -1,42 +1,10 @@
-import 'package:dongsoop/domain/board/recruit/entities/recruit_list_entity.dart';
 import 'package:dongsoop/domain/board/recruit/enum/recruit_type.dart';
 import 'package:dongsoop/domain/board/recruit/use_cases/recruit_list_use_case.dart';
 import 'package:dongsoop/presentation/board/providers/recruit/recruit_list_use_case_provider.dart';
+import 'package:dongsoop/presentation/board/recruit/list/state/recruit_list_state.dart';
 import 'package:riverpod_annotation/riverpod_annotation.dart';
 
 part 'recruit_list_view_model.g.dart';
-
-class RecruitListState {
-  final List<RecruitListEntity> posts;
-  final bool isLoading;
-  final bool hasMore;
-  final String? error;
-  final int page;
-
-  RecruitListState({
-    this.posts = const [],
-    this.isLoading = false,
-    this.hasMore = true,
-    this.error,
-    this.page = 0,
-  });
-
-  RecruitListState copyWith({
-    List<RecruitListEntity>? posts,
-    bool? isLoading,
-    bool? hasMore,
-    String? error,
-    int? page,
-  }) {
-    return RecruitListState(
-      posts: posts ?? this.posts,
-      isLoading: isLoading ?? this.isLoading,
-      hasMore: hasMore ?? this.hasMore,
-      error: error,
-      page: page ?? this.page,
-    );
-  }
-}
 
 @riverpod
 class RecruitListViewModel extends _$RecruitListViewModel {
@@ -45,6 +13,7 @@ class RecruitListViewModel extends _$RecruitListViewModel {
   late String _departmentCode;
 
   static const int _pageSize = 10;
+  bool _initialized = false;
 
   @override
   RecruitListState build({
@@ -55,12 +24,17 @@ class RecruitListViewModel extends _$RecruitListViewModel {
     _type = type;
     _departmentCode = departmentCode;
 
+    // 상태 초기화
+    state = RecruitListState();
+
     Future.microtask(() => _initialize());
 
-    return RecruitListState();
+    return state;
   }
 
   Future<void> _initialize() async {
+    if (_initialized) return;
+    _initialized = true;
     await loadNextPage();
   }
 
@@ -78,8 +52,16 @@ class RecruitListViewModel extends _$RecruitListViewModel {
 
       final filtered = newPosts.where((e) => e.state).toList();
 
+      // 중복 제거 (id 기준)
+      final uniquePosts = filtered
+          .where(
+            (newPost) =>
+                !state.posts.any((existing) => existing.id == newPost.id),
+          )
+          .toList();
+
       state = state.copyWith(
-        posts: [...state.posts, ...filtered],
+        posts: [...state.posts, ...uniquePosts],
         page: state.page + 1,
         isLoading: false,
         hasMore: filtered.length == _pageSize,
@@ -91,7 +73,8 @@ class RecruitListViewModel extends _$RecruitListViewModel {
   }
 
   Future<void> refresh() async {
+    _initialized = false;
     state = RecruitListState();
-    await loadNextPage();
+    await _initialize();
   }
 }
